feat: server setup
Initial server setup with the following configurations:

1. Create handler that will handle request and response cycle from Next
application instance.
2. Create instance of Sentiment that will handle analysis of Sentiments
3. Pusher instance with the environment variables setup from
initialization of DefinePlugin in the next.config.js
4. Prepare the application:
 i. This will setup and initialize an express application instance
adding the following middleware
   a. Adding Cross Origin Resource Sharing to allow sharing of resources
from different origins
   b. bodyParser will parse the responses and requests to JSON for
consumption in the application.
   c. Adds the handler middleware to handle get requests on all urls
   d. Start up the server instance to listen to requests

This also adds a logger server that is used for logging. This will track
errors and information and store the logs in files, locatted under the
log directory, which winston will create.