feat: layout wrapper
Create Layout wrapper that will be used as the common layout for all
pages in the application. This will be used by the application to set a
default Head component that will have meta data for the application.

This is a basic stateless React component that is passed to props
1. children : which is any of the types, object, element, array, node,
element
2. pageTitle: which is the title of the page