feat: chat
Create Chat component that will handle the chat messaging and will be
used to open and bind to the chat-room channel and subscribe to
new-message event.

This will then update the state of the chats array in the component and
will re-render on new chat messages. This will then bind to the
connection event on the pusher instance and will then update the state
of the component